moved sprite into sprites namespace
